1.6 VBox working but still some make/docker build issues

+++ packer/voltha/voltha_compose_stop.sh
@@ -2,12 +2,32 @@
 docker stop voltha-onu_mock
 docker rm voltha-onu_mock
 
-docker-compose -f /home/voltha/compose/docker-compose-min-system-test.yml stop
-docker-compose -f /home/voltha/compose/docker-compose-local-auth-test.yml stop
+if [ -f /home/voltha/compose/docker-compose-min-system-test.yml ]
+then
+    docker-compose -f /home/voltha/compose/docker-compose-min-system-test.yml stop
+fi
+if [ -f /home/voltha/compose/docker-compose-system-test.yml ]
+then
+    docker-compose -f /home/voltha/compose/docker-compose-system-test.yml stop
+fi
+if [ -f /home/voltha/compose/docker-compose-local-auth-test.yml ]
+then
+    docker-compose -f /home/voltha/compose/docker-compose-local-auth-test.yml stop
+fi
 
-docker-compose -f /home/voltha/compose/docker-compose-min-system-test.yml rm -f
-docker-compose -f /home/voltha/compose/docker-compose-local-auth-test.yml rm -f
 
+if [ -f /home/voltha/compose/docker-compose-min-system-test.yml ]
+then
+    docker-compose -f /home/voltha/compose/docker-compose-min-system-test.yml rm -f
+fi
+if [ -f /home/voltha/compose/docker-compose-system-test.yml ]
+then
+    docker-compose -f /home/voltha/compose/docker-compose-system-test.yml rm -f
+fi
+if [ -f /home/voltha/compose/docker-compose-local-auth-test.yml ]
+then
+    docker-compose -f /home/voltha/compose/docker-compose-local-auth-test.yml rm -f
+fi
 #
 # TODO: Need to add support for the following
 #       - Move  mock into a compose file
